+++ app/src/main/java/com/edu6tron/spiritualcompanion/nativepreview/media/NativeDevotionalPlayer.kt
@@ -97,6 +97,13 @@ class NativeDevotionalPlayer @Inject constructor(
     )
   }
 
+  fun playOfflineSoundscape(soundscape: OfflineSoundscape) {
+    play(
+      "android.resource://${appContext.packageName}/${soundscape.resourceId}",
+      soundscape.playbackLabel,
+    )
+  }
+
   fun stop() {
     progressHandler.removeCallbacks(progressTick)
     runCatching { player.stop() }